move test for base classes to the right directory
