+++ tests/codes/eurocode/en_1992_1_1_2004/chapter_4_durability_and_cover/_base_classes/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing all tests for base classes for EN 1992-1-1:2004: Chapter 4 - Durability and cover to reinforcement."""
