+++ tests/codes/eurocode/en_1992_1_1_2004/chapter_4_durability_and_cover/_base_classes/test_exposure_classes.py
